added rdf_collection
I added a liquid filter that can traverse through a given rdf_resource if it is a container type.
If this condition is met, it will return all possible container elements, regardless of any semantic conditions met (or not met) by the container in question.
Additionally, I added a few layouts to demonstrate the usage of collections and container filters.